+++ qoqo_qasm/src/lib.rs
@@ -11,19 +11,31 @@
 // limitations under the License.
 
 #![deny(missing_docs)]
-#![warn(rustdoc::private_intra_doc_links)]
-#![warn(rustdoc::missing_crate_level_docs)]
-#![warn(rustdoc::missing_doc_code_examples)]
-#![warn(rustdoc::private_doc_tests)]
+#![warn(private_intra_doc_links)]
+#![warn(missing_crate_level_docs)]
+#![warn(missing_doc_code_examples)]
+#![warn(private_doc_tests)]
 #![deny(missing_debug_implementations)]
 
-//! # roqoqo-qasm
+//! # qoqo-qasm
 //!
 //! QASM interface for qoqo.
 //!
 //! Translates qoqo operations and circuits to QASM operations via the interface, and Create a Qasm file with QasmBackend.
 
+use pyo3::prelude::*;
+use pyo3::wrap_pyfunction;
+
 mod backend;
 pub use backend::*;
+
 mod interface;
 pub use interface::*;
+
+#[pymodule]
+fn qoqo_qasm(_py: Python, module: &PyModule) -> PyResult<()> {
+    module.add_class::<QasmBackendWrapper>()?;
+    module.add_function(wrap_pyfunction!(qasm_call_circuit, module)?)?;
+    module.add_function(wrap_pyfunction!(qasm_call_operation, module)?)?;
+    Ok(())
+}