+++ roqoqo-qasm/src/backend.rs
@@ -1,4 +1,4 @@
-// Copyright © 2021 HQS Quantum Simulations GmbH. All Rights Reserved.
+// Copyright © 2021-2023 HQS Quantum Simulations GmbH. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 // in compliance with the License. You may obtain a copy of the License at
@@ -10,7 +10,8 @@
 // express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::call_operation;
+use crate::{call_operation, gate_definition};
+use qoqo_calculator::CalculatorFloat;
 use roqoqo::operations::*;
 use roqoqo::{Circuit, RoqoqoBackendError};
 use std::fs::File;
@@ -47,8 +48,6 @@ impl Backend {
     /// # Arguments
     ///
     /// * `qubit_register_name` - The number of qubits in the backend.
-    /// * ``
-    ///
     pub fn new(qubit_register_name: Option<String>) -> Self {
         match qubit_register_name {
             None => Self {
@@ -59,9 +58,6 @@ impl Backend {
             },
         }
     }
-}
-
-impl Backend {
     /// Translates an iterator over operations to a valid QASM string.
     ///
     ///
@@ -77,10 +73,38 @@ impl Backend {
         &self,
         circuit: impl Iterator<Item = &'a Operation>,
     ) -> Result<String, RoqoqoBackendError> {
+        let mut definitions: String = "".to_string();
         let mut data: String = "".to_string();
-        let mut qasm_string = String::from("OPENQASM 2.0;\ninclude \"qelib1.inc\";\n\n");
+        let mut qasm_string = String::from("OPENQASM 2.0;\n\n");
 
         let mut number_qubits_required: usize = 0;
+        let mut already_seen_definitions: Vec<String> = vec![
+            "RotateX".to_string(),
+            "RotateY".to_string(),
+            "RotateZ".to_string(),
+            "CNOT".to_string(),
+        ];
+        definitions.push_str("gate u3(theta,phi,lambda) q { U(theta,phi,lambda) q; }\n");
+        definitions.push_str("gate u2(phi,lambda) q { U(pi/2,phi,lambda) q; }\n");
+        definitions.push_str("gate u1(lambda) q { U(0,0,lambda) q; }\n");
+        definitions.push_str(&gate_definition(&Operation::from(RotateX::new(
+            0,
+            CalculatorFloat::from(0.0),
+        )))?);
+        definitions.push('\n');
+        definitions.push_str(&gate_definition(&Operation::from(RotateY::new(
+            0,
+            CalculatorFloat::from(0.0),
+        )))?);
+        definitions.push('\n');
+        definitions.push_str(&gate_definition(&Operation::from(RotateZ::new(
+            0,
+            CalculatorFloat::from(0.0),
+        )))?);
+        definitions.push('\n');
+        definitions.push_str(&gate_definition(&Operation::from(CNOT::new(0, 1)))?);
+        definitions.push('\n');
+
         for op in circuit {
             if let InvolvedQubits::Set(involved_qubits) = op.involved_qubits() {
                 number_qubits_required =
@@ -89,9 +113,20 @@ impl Backend {
                         Some(n) => *n,
                     })
             }
+            if !already_seen_definitions.contains(&op.hqslang().to_string()) {
+                already_seen_definitions.push(op.hqslang().to_string());
+                definitions.push_str(&gate_definition(op)?);
+                if !definitions.is_empty() {
+                    definitions.push('\n');
+                }
+            }
             data.push_str(&call_operation(op, &self.qubit_register_name)?);
-            data.push('\n');
+            if !data.is_empty() {
+                data.push('\n');
+            }
         }
+        qasm_string.push_str(definitions.as_str());
+
         qasm_string.push_str(
             format!(
                 "qreg {}[{}];\n",
@@ -100,7 +135,6 @@ impl Backend {
             )
             .as_str(),
         );
-
         qasm_string.push_str(data.as_str());
 
         Ok(qasm_string)